type check

+++ src/main/java/xyz/acrylicstyle/doubletimecommandsbungee/types/Ban.java
@@ -6,10 +6,10 @@ public class Ban {
     private final int id;
     private final UUID player;
     private final String reason;
-    private final int expires;
+    private final long expires;
     private final UUID executor;
 
-    public Ban(int id, UUID player, String reason, int expires, UUID executor) {
+    public Ban(int id, UUID player, String reason, long expires, UUID executor) {
         this.id = id;
         this.player = player;
         this.reason = reason;
@@ -20,6 +20,6 @@ public class Ban {
     public final int getBanId() { return this.id; }
     public final UUID getPlayer() { return this.player; }
     public final String getReason() { return this.reason == null ? "None" : this.reason; }
-    public final int getExpires() { return this.expires; }
+    public final long getExpires() { return this.expires; }
     public final UUID getExecutor() { return this.executor; }
 }
